fix for issue where offline order import would not associate the customer with the order (could not find customer by ID)
