+++ src/Feature/Orders/engine/Pipelines/Arguments/Store.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Sitecore.HabitatHome.Feature.Orders.Engine.Pipelines.Arguments
+{
+    public class Store
+    {
+        public string Name { get; set; }
+
+        public string Address { get; set; }
+
+        public string City { get; set; }
+
+        public string State { get; set; }
+
+        public string ZipCode { get; set; }
+
+        public string Country { get; set; }
+    }
+}
