+++ src/Feature/Orders/engine/SampleData/PostmanOfflineOrder.json
@@ -3,6 +3,7 @@
     "OrderPlacedDate": "2018-02-14T22:32:41.5332063+00:00",
     "OrderConfirmationId": "610003123232",
     "Email": "prasa@gmail.com",
+    "Domain": "CommerceUsers", 
     "Language": "en-US",
     "Status": "StoreOrder",
     "Lines": [
