+++ src/Feature/Orders/engine/SampleData/SampleUploadSingleOrder.json
@@ -4,6 +4,7 @@
       "OrderPlacedDate": "2018-02-14T22:32:41.5332063+00:00",
       "OrderConfirmationId": "509",
       "Email": "prasad@gmail.com",
+      "Domain": "CommerceUsers",
       "Language": "en-US",
       "Status": "StoreOrder",
       "Lines": [
